+++ app/views/optIn/BeforeYouStart.scala.html
@@ -16,9 +16,8 @@
 
 @import uk.gov.hmrc.govukfrontend.views.html.components._
 @import views.html.components._
-@import views.html.layouts.unifiedLayout
-@import models.incomeSourceDetails.TaxYear
 @import views.html.helpers.injected.ErrorSummary
+@import views.html.layouts.unifiedLayout
 
 @this(
         mainTemplate: unifiedLayout,
