+++ app/views/optOut/OptOutChooseTaxYear.scala.html
@@ -18,7 +18,7 @@
 @import views.html.components._
 @import views.html.layouts.unifiedLayout
 @import models.incomeSourceDetails.TaxYear
-@import services.optout.OptOutService.QuarterlyUpdatesCountForTaxYearModel
+@import services.reportingfreq.ReportingFrequency.QuarterlyUpdatesCountForTaxYearModel
 @import views.html.helpers.injected.ErrorSummary
 @import forms.optOut.ConfirmOptOutMultiTaxYearChoiceForm
